Added process_search_parameters(), get_valid_sort_for_resource() into search_params file from resources file
